backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# NLP & AI
spacy==3.7.2
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP Client
httpx==0.26.0
